TokenStaking
The TokenStaking contract is a simple contract that allows users to stake their tokens and receive rewards for doing so. By staking tokens, users can earn a reward rate (expressed in decimal, for example 0.1 for a 10% reward rate) based on the amount of tokens staked. The contract also allows users to claim the rewards they have earned. We hope you will use it in your projects.